Refactor: Split scope module
`GlobalScope` goes into the compile module, as it will become a local
type in the future. `GlobalScope` keeps track of information that is
needed during compilation. The generation of the abstract syntax tree,
which we will implement in a future PR, will use a different scope.

Rename `scope.rs` to `pattern.rs`. In the next commit, we will move the
remaining pattern logic into this module.
